refactor(core): include `svgdom` in `core`
No longer a separate executable, but something that we build as part of our core library.
If we ever refactor our core library into smaller pieces, this is natural to split back out.

The unit tests are still a no-op, but they're something Dave should be able to work on later.
