+++ core/CMakeLists.txt
@@ -1,7 +1,7 @@
 # Build this library statically, because it will be consumed in two different
 # SOs The first is the actual pepplib library, and the other is nanobind.
-file(GLOB_RECURSE sources CONFIGURE_DEPENDS "${CMAKE_CURRENT_LIST_DIR}/core/*.cpp"
-     "${CMAKE_CURRENT_LIST_DIR}/core/*.hpp")
+file(GLOB_RECURSE sources CONFIGURE_DEPENDS "${CMAKE_CURRENT_LIST_DIR}/*.cpp"
+     "${CMAKE_CURRENT_LIST_DIR}/*.hpp")
 add_library(pepp-core STATIC ${sources})
 set_target_properties(pepp-core PROPERTIES POSITION_INDEPENDENT_CODE ON)
 target_link_libraries(
@@ -15,8 +15,7 @@ target_link_libraries(
          elfio
          zpp_bits
          cnl
-        rapidxml
-)
+         rapidxml)
 target_include_directories(pepp-core PUBLIC ${CMAKE_CURRENT_LIST_DIR})
 
 if(PEPP_PRECOMPILE_HEADERS)
@@ -36,4 +35,3 @@ if(PEPP_PRECOMPILE_HEADERS)
   pepp_precompile_headers(
     pepp-core PRIVATE "${CMAKE_CURRENT_LIST_DIR}/core/math/geom/interval.hpp")
 endif()
-add_subdirectory(svgdom)
