--- core/svgdom/.gitignore
@@ -1,15 +1,0 @@
-### Windows
-# Windows thumbnail cache files
-Thumbs.db
-
-# Folder config file
-[Dd]esktop.ini
-
-# Recycle Bin used on file shares
-$RECYCLE.BIN/
-
-# Windows shortcuts
-*.lnk
-
-# Qtcreator directory
-.qtcreator
