--- core/svgdom/src/SvgCommentElement.hpp
@@ -1,22 +1,0 @@
-#pragma once
-
-#include <string>
-
-#include "SvgElement.hpp"
-
-class SvgCommentElement final : public Cloneable<SvgCommentElement, SvgElement>
-{
-public:
-    SvgCommentElement();
-    explicit SvgCommentElement(const std::string &comment);
-    ~SvgCommentElement() = default;
-    //Can copy and move
-    SvgCommentElement(const SvgCommentElement &) = default;
-    SvgCommentElement &operator=(const SvgCommentElement &) = default;
-    SvgCommentElement(SvgCommentElement &&) noexcept = default;
-    SvgCommentElement &operator=(SvgCommentElement &&) noexcept = default;
-
-    //  Called by base class
-    void toXml(SvgRope &output) const override;
-    bool setAttribute(const std::string &key, const std::string &value) override;
-};
