--- core/svgdom/src/SvgType.hpp
@@ -1,34 +1,0 @@
-#pragma once
-
-class SvgType
-{
-public:
-    enum class Type {
-        SvgUnknownElement = 0,
-        //  Dom elements
-        SvgBasicElement,
-        SvgCommentElement,
-        SvgCDataElement,
-        //  SvgSpecific elements
-        SvgElementParent = 0x0100,
-        SvgDescElement,
-        SvgMetadataElement,
-        SvgStyleElement,
-        SvgTitleElement,
-        SvgGraphicElement = 0x8000,
-        SvgSvgElement,
-        //  Graphic elements
-        SvgGeometry = 0x8100,
-        SvgGElement,
-        SvgDefsElement,
-        SvgSymbolElement,
-        SvgUseElement,
-        SvgSwitchElement,
-        SvgCircleElement,
-        SvgLineElement,
-        SvgPathElement,
-        SvgPolygonElement,
-        SvgRectElement,
-        SvgTextElement,
-    };
-};
